+++ src/main/kotlin/app/revanced/cli/command/ListPatchesCommand.kt
@@ -9,41 +9,41 @@ import java.io.File
 import java.util.logging.Logger
 
 
-@Command(name = "list-patches", description = ["List patches from supplied patch bundles"])
+@Command(name = "list-patches", description = ["List patches from supplied patch bundles."])
 internal object ListPatchesCommand : Runnable {
     private val logger = Logger.getLogger(ListPatchesCommand::class.java.name)
 
     @Parameters(
-        description = ["Paths to patch bundles"], arity = "1..*"
+        description = ["Paths to patch bundles."], arity = "1..*"
     )
     private lateinit var patchBundles: Array<File>
 
     @Option(
-        names = ["-d", "--with-descriptions"], description = ["List their descriptions"], showDefaultValue = ALWAYS
+        names = ["-d", "--with-descriptions"], description = ["List their descriptions."], showDefaultValue = ALWAYS
     )
     private var withDescriptions: Boolean = true
 
     @Option(
         names = ["-p", "--with-packages"],
-        description = ["List the packages the patches are compatible with"],
+        description = ["List the packages the patches are compatible with."],
         showDefaultValue = ALWAYS
     )
     private var withPackages: Boolean = false
 
     @Option(
         names = ["-v", "--with-versions"],
-        description = ["List the versions of the apps the patches are compatible with"],
+        description = ["List the versions of the apps the patches are compatible with."],
         showDefaultValue = ALWAYS
     )
     private var withVersions: Boolean = false
 
     @Option(
-        names = ["-o", "--with-options"], description = ["List the options of the patches"], showDefaultValue = ALWAYS
+        names = ["-o", "--with-options"], description = ["List the options of the patches."], showDefaultValue = ALWAYS
     )
     private var withOptions: Boolean = false
 
     @Option(
-        names = ["-f", "--filter-package-name"], description = ["Filter patches by package name"]
+        names = ["-f", "--filter-package-name"], description = ["Filter patches by package name."]
     )
     private var packageName: String? = null